Fixed os classes bugs under Windows.

+++ src/win32/Mouse.cpp
@@ -1,8 +1,10 @@
+#include <Windows.h>
+
 #include "Mouse.hpp"
 
 using namespace air::os;
 
 void		Mouse::setPosition(double x, double y)
 {
-  SetCursorPos(x, y);
+  SetCursorPos((int) x, (int) y);
 }
